+++ Sources/StreamChatUI/Appearance+Images.swift
@@ -54,9 +54,7 @@ public extension Appearance {
         public var shrinkInputArrow: UIImage = loadImageSafely(with: "arrow_shrink_input")
         public var sendArrow: UIImage = loadImageSafely(with: "arrow_send")
         public var scrollDownArrow: UIImage = loadImageSafely(with: "arrow_down")
-        public var messageSent: UIImage = loadImageSafely(with: "checkmark_grey")
         public var whiteCheckmark: UIImage = loadImageSafely(with: "checkmark_white")
-        public var readByAll: UIImage = loadImageSafely(with: "checkmark_double")
         public var confirmCheckmark: UIImage = loadImageSafely(with: "checkmark_confirm")
         public var bigConfirmCheckmark: UIImage = loadImageSafely(with: "checkmark_confirm_big")
         public var folder: UIImage = loadImageSafely(with: "folder")
@@ -69,6 +67,12 @@ public extension Appearance {
             }
         }()
 
+        // MARK: - Message Receipts
+        
+        public var messageDeliveryStatusSending: UIImage = loadImageSafely(with: "message_receipt_sending")
+        public var messageDeliveryStatusSent: UIImage = loadImageSafely(with: "message_receipt_sent")
+        public var messageDeliveryStatusRead: UIImage = loadImageSafely(with: "message_receipt_read")
+        
         // MARK: - Reactions
 
         public var reactionLoveSmall: UIImage = loadImageSafely(with: "reaction_love_small")