+++ Sources/StreamChatUI/ChatMessageList/ChatMessage/ChatMessageContentView.swift
@@ -51,8 +51,14 @@ open class ChatMessageContentView: _View, ThemeProvider {
     /// The delegate responsible for action handling.
     public weak var delegate: ChatMessageContentViewDelegate?
 
+    // TODO: Aggregate message and channel under one `struct Content` roof in v5
     /// The message this view displays.
     open var content: ChatMessage? {
+        didSet { updateContentIfNeeded() }
+    }
+    
+    /// The channel the message is sent to.
+    open var channel: ChatChannel? {
         didSet { updateContentIfNeeded() }
     }
 
@@ -131,6 +137,11 @@ open class ChatMessageContentView: _View, ThemeProvider {
     /// Exists if `layout(options: MessageLayoutOptions)` was invoked with the options containing `.threadInfo`.
     public private(set) var threadArrowView: ChatThreadArrowView?
 
+    /// Shows message delivery status.
+    /// Exists if `layout(options: ChatMessageLayoutOption)` was invoked with the options
+    /// containing `.messageDeliveryStatus`.
+    public private(set) var deliveryStatusView: ChatMessageDeliveryStatusView?
+    
     /// An object responsible for injecting the views needed to display the attachments content.
     public private(set) var attachmentViewInjector: AttachmentViewInjector?
 
@@ -287,6 +298,9 @@ open class ChatMessageContentView: _View, ThemeProvider {
                 onlyVisibleToYouContainer!.addArrangedSubview(createOnlyVisibleToYouImageView())
                 onlyVisibleToYouContainer!.addArrangedSubview(createOnlyVisibleToYouLabel())
                 footnoteSubviews.append(onlyVisibleToYouContainer!)
+            }
+            if options.contains(.deliveryStatusIndicator) {
+                footnoteSubviews.append(createDeliveryStatusView())
             }
             if attachmentViewInjector?.fillAllAvailableWidth == true {
                 footnoteSubviews.append(.spacer(axis: .horizontal))
@@ -551,6 +565,12 @@ open class ChatMessageContentView: _View, ThemeProvider {
                 didTapOnReaction: nil
             )
         }
+        
+        // Delivery status
+        deliveryStatusView?.content = {
+            guard let channel = channel, let message = content else { return nil }
+            return .init(message: message, channel: channel)
+        }()
     }
 
     override open func tintColorDidChange() {
@@ -820,6 +840,18 @@ open class ChatMessageContentView: _View, ThemeProvider {
         }
         return onlyVisibleToYouLabel!
     }
+    
+    /// Instantiates, configures and assigns `deliveryStatusView` when called for the first time.
+    /// - Returns: The `deliveryStatusView` subview.
+    open func createDeliveryStatusView() -> ChatMessageDeliveryStatusView {
+        if deliveryStatusView == nil {
+            deliveryStatusView = components
+                .messageDeliveryStatusView
+                .init()
+                .withAccessibilityIdentifier(identifier: "deliveryStatusView")
+        }
+        return deliveryStatusView!
+    }
 }
 
 private extension ChatMessage {
@@ -838,7 +870,8 @@ private extension ChatMessageLayoutOptions {
     static let footnote: Self = [
         .onlyVisibleToYouIndicator,
         .authorName,
-        .timestamp
+        .timestamp,
+        .deliveryStatusIndicator
     ]
     
     var hasFootnoteOptions: Bool {