+++ Sources/StreamChatUI/ChatMessageList/ChatMessage/ChatMessageLayoutOptionsResolver.swift
@@ -103,6 +103,9 @@ open class ChatMessageLayoutOptionsResolver {
         if message.isLastActionFailed {
             options.insert(.errorIndicator)
         }
+        if isLastInSequence && canShowDeliveryStatus(for: message, in: channel) {
+            options.insert(.deliveryStatusIndicator)
+        }
 
         return options
     }
@@ -205,6 +208,29 @@ open class ChatMessageLayoutOptionsResolver {
             return config.deletedMessagesVisibility == .visibleForCurrentUser
         default:
             return false
+        }
+    }
+    
+    func canShowDeliveryStatus(for message: ChatMessage, in channel: ChatChannel) -> Bool {
+        guard message.isSentByCurrentUser else {
+            // Delivery status can only be shown for a message sent by the current user.
+            return false
+        }
+        
+        guard message.type != .deleted, message.type != .ephemeral else {
+            // Delivery status should not be shown for deleted & ephemeral messages.
+            return false
+        }
+        
+        switch message.localState {
+        case .pendingSend, .sending, .pendingSync, .syncing, .deleting:
+            // Delivery status should always be shown for message in `pending` state.
+            return true
+        case .sendingFailed, .syncingFailed, .deletingFailed:
+            // Delivery status should not be shown for failed message (the error indicator is shown instead).
+            return false
+        case nil:
+            return channel.config.readEventsEnabled
         }
     }
 }